+++ src/main/java/com/backend/server/model/repository/rateLimit/RedisRateLimitRepository.java
@@ -2,14 +2,12 @@ package com.backend.server.model.repository.rateLimit;
 
 import com.backend.server.api.common.exceptionHandler.exception.TooManyRequestException;
 import com.backend.server.config.RedisScriptConfig;
-import com.backend.server.config.annotation.rateLimit.LimitRequestPerTime;
+import com.backend.server.support.rateLimit.LimitRequestPerTime;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.Collections;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -17,18 +15,13 @@ import org.springframework.stereotype.Component;
 public class RedisRateLimitRepository implements RateLimitRepository {
 
     private final RedisScriptConfig scripts;
-    private final RedisTemplate<String, String> redisTemplate;
 
     @Override
     public Long add(String key, LimitRequestPerTime limit) throws TooManyRequestException {
         final long windowMillis = limit.timeUnit().toMillis(limit.time());
         final int threshold = limit.count();
 
-        Long result = redisTemplate.execute(
-                scripts.getRateLimit_add(),
-                Collections.singletonList(key),
-                String.valueOf(windowMillis),
-                String.valueOf(threshold));
+        Long result = scripts.rateLimitAdd(key, windowMillis, threshold);
 
         // 요청 횟수 제한을 초과했을때
         if (result < 0) {
@@ -46,8 +39,6 @@ public class RedisRateLimitRepository implements RateLimitRepository {
 
     @Override
     public Long delete(String key) {
-        return redisTemplate.execute(
-                scripts.getRateLimit_delete(),
-                Collections.singletonList(key));
+        return scripts.rateLimitDelete(key);
     }
 }
