+++ src/test/java/com/backend/server/api/admin/classroom/controller/AdminClassroomControllerTest.java
@@ -3,9 +3,9 @@ package com.backend.server.api.admin.classroom.controller;
 import static com.backend.server.fixture.ClassroomFixture.강의실1;
 import static com.backend.server.fixture.ClassroomFixture.강의실2;
 import static com.backend.server.fixture.UserFixture.관리자1;
-import static com.backend.server.util.MockMvcUtil.convertToJson;
-import static com.backend.server.util.MockMvcUtil.convertToParams;
-import static com.backend.server.util.MockMvcUtil.jsonPathEquals;
+import static com.backend.server.support.MockMvcUtil.convertToJson;
+import static com.backend.server.support.MockMvcUtil.convertToParams;
+import static com.backend.server.support.MockMvcUtil.jsonPathEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
