+++ src/test/java/com/backend/server/api/admin/equipment/controller/AdminEquipmentModelControllerTest.java
@@ -6,7 +6,7 @@ import com.backend.server.model.entity.equipment.EquipmentCategory;
 import com.backend.server.model.entity.equipment.EquipmentModel;
 import com.backend.server.model.repository.equipment.EquipmentCategoryRepository;
 import com.backend.server.model.repository.equipment.EquipmentModelRepository;
-import com.backend.server.util.MockMvcUtil;
+import com.backend.server.support.MockMvcUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
