+++ src/test/java/com/backend/server/api/common/s3/controller/CommonPresignedUrlControllerTest.java
@@ -1,12 +1,12 @@
 package com.backend.server.api.common.s3.controller;
 
-import static com.backend.server.util.MockMvcUtil.toJsonPathDocument;
+import static com.backend.server.support.MockMvcUtil.toJsonPathDocument;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.backend.server.config.ControllerTest;
-import com.backend.server.util.S3ApiUtil;
+import com.backend.server.support.S3ApiUtil;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
